refactor: naming + improved SpectralResolver
